chore: rename storev2 to legacy_store
