+++ waku/v2/node/wakuoptions_test.go
@@ -1,13 +1,15 @@
 package node
 
 import (
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/waku-org/go-waku/waku/v2/protocol/pb"
-	r "github.com/waku-org/go-zerokit-rln/rln"
-	"go.uber.org/zap"
 	"net"
 	"testing"
 	"time"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/waku-org/go-waku/waku/v2/protocol/legacy_store"
+	"github.com/waku-org/go-waku/waku/v2/protocol/pb"
+	r "github.com/waku-org/go-zerokit-rln/rln"
+	"go.uber.org/zap"
 
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/multiformats/go-multiaddr"
@@ -15,7 +17,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/waku-org/go-waku/tests"
 	"github.com/waku-org/go-waku/waku/persistence"
-	"github.com/waku-org/go-waku/waku/v2/protocol/store"
 )
 
 func handleSpam(msg *pb.WakuMessage, topic string) error {
@@ -41,8 +42,8 @@ func TestWakuOptions(t *testing.T) {
 	addr, err := multiaddr.NewMultiaddr("/ip4/0.0.0.0/tcp/4000/ws")
 	require.NoError(t, err)
 
-	storeFactory := func(w *WakuNode) store.Store {
-		return store.NewWakuStore(w.opts.messageProvider, w.peermanager, w.timesource, prometheus.DefaultRegisterer, w.log)
+	storeFactory := func(w *WakuNode) legacy_store.Store {
+		return legacy_store.NewWakuStore(w.opts.messageProvider, w.peermanager, w.timesource, prometheus.DefaultRegisterer, w.log)
 	}
 
 	options := []WakuNodeOption{
@@ -87,8 +88,8 @@ func TestWakuRLNOptions(t *testing.T) {
 	addr, err := multiaddr.NewMultiaddr("/ip4/0.0.0.0/tcp/4000/ws")
 	require.NoError(t, err)
 
-	storeFactory := func(w *WakuNode) store.Store {
-		return store.NewWakuStore(w.opts.messageProvider, w.peermanager, w.timesource, prometheus.DefaultRegisterer, w.log)
+	storeFactory := func(w *WakuNode) legacy_store.Store {
+		return legacy_store.NewWakuStore(w.opts.messageProvider, w.peermanager, w.timesource, prometheus.DefaultRegisterer, w.log)
 	}
 
 	index := r.MembershipIndex(5)
