+++ waku/v2/protocol/legacy_store/utils_test.go
@@ -1,4 +1,4 @@
-package store
+package legacy_store
 
 import (
 	"database/sql"
