+++ waku/v2/protocol/legacy_store/waku_resume_test.go
@@ -1,4 +1,4 @@
-package store
+package legacy_store
 
 import (
 	"context"
