+++ waku/v2/protocol/legacy_store/waku_store_client.go
@@ -1,4 +1,4 @@
-package store
+package legacy_store
 
 import (
 	"context"
@@ -15,8 +15,8 @@ import (
 	"github.com/waku-org/go-waku/waku/v2/peermanager"
 	"github.com/waku-org/go-waku/waku/v2/peerstore"
 	"github.com/waku-org/go-waku/waku/v2/protocol"
+	"github.com/waku-org/go-waku/waku/v2/protocol/legacy_store/pb"
 	wpb "github.com/waku-org/go-waku/waku/v2/protocol/pb"
-	"github.com/waku-org/go-waku/waku/v2/protocol/store/pb"
 )
 
 type Query struct {
