+++ waku/v2/protocol/legacy_store/waku_store_common.go
@@ -1,4 +1,4 @@
-package store
+package legacy_store
 
 import (
 	"context"
