+++ waku/v2/protocol/store/pb/validation.go
@@ -4,44 +4,75 @@ import (
 	"errors"
 )
 
-// MaxContentFilters is the maximum number of allowed content filters in a query
-const MaxContentFilters = 10
+// MaxContentTopics is the maximum number of allowed contenttopics in a query
+const MaxContentTopics = 10
 
 var (
-	errMissingRequestID   = errors.New("missing RequestId field")
-	errMissingQuery       = errors.New("missing Query field")
-	errRequestIDMismatch  = errors.New("requestID in response does not match request")
-	errMaxContentFilters  = errors.New("exceeds the maximum number of content filters allowed")
-	errEmptyContentTopics = errors.New("one or more content topics specified is empty")
+	errMissingRequestID       = errors.New("missing RequestId field")
+	errMessageHashOtherFields = errors.New("cannot use MessageHashes with ContentTopics/PubsubTopic")
+	errRequestIDMismatch      = errors.New("requestID in response does not match request")
+	errMaxContentTopics       = errors.New("exceeds the maximum number of ContentTopics allowed")
+	errEmptyContentTopic      = errors.New("one or more content topics specified is empty")
+	errMissingPubsubTopic     = errors.New("missing PubsubTopic field")
+	errMissingContentTopics   = errors.New("missing ContentTopics field")
+	errMissingStatusCode      = errors.New("missing StatusCode field")
+	errMissingMessage         = errors.New("missing Message field")
+	errInvalidTimeRange       = errors.New("invalid time range")
+	errInvalidMessageHash     = errors.New("invalid message hash")
 )
 
-func (x *HistoryQuery) Validate() error {
-	if len(x.ContentFilters) > MaxContentFilters {
-		return errMaxContentFilters
-	}
-
-	for _, m := range x.ContentFilters {
-		if m.ContentTopic == "" {
-			return errEmptyContentTopics
-		}
-	}
-
-	return nil
-}
-
-func (x *HistoryRPC) ValidateQuery() error {
+func (x *StoreQueryRequest) Validate() error {
 	if x.RequestId == "" {
 		return errMissingRequestID
 	}
 
-	if x.Query == nil {
-		return errMissingQuery
-	}
+	if len(x.MessageHashes) != 0 {
+		if len(x.ContentTopics) != 0 || x.GetPubsubTopic() != "" {
+			return errMessageHashOtherFields
+		}
 
-	return x.Query.Validate()
+		for _, x := range x.MessageHashes {
+			if len(x) != 32 {
+				return errInvalidMessageHash
+			}
+		}
+	} else {
+		if x.GetPubsubTopic() == "" {
+			return errMissingPubsubTopic
+		}
+
+		if len(x.ContentTopics) == 0 {
+			return errMissingContentTopics
+		} else if len(x.ContentTopics) > MaxContentTopics {
+			return errMaxContentTopics
+		} else {
+			for _, m := range x.ContentTopics {
+				if m == "" {
+					return errEmptyContentTopic
+				}
+			}
+		}
+
+		if x.GetTimeStart() > 0 && x.GetTimeEnd() > 0 && x.GetTimeStart() > x.GetTimeEnd() {
+			return errInvalidTimeRange
+		}
+	}
+	return nil
 }
 
-func (x *HistoryResponse) Validate() error {
+func (x *StoreQueryResponse) Validate(requestID string) error {
+	if x.RequestId == "" {
+		return errMissingRequestID
+	}
+
+	if x.RequestId != requestID {
+		return errRequestIDMismatch
+	}
+
+	if x.StatusCode == nil {
+		return errMissingStatusCode
+	}
+
 	for _, m := range x.Messages {
 		if err := m.Validate(); err != nil {
 			return err
@@ -51,18 +82,14 @@ func (x *HistoryResponse) Validate() error {
 	return nil
 }
 
-func (x *HistoryRPC) ValidateResponse(requestID string) error {
-	if x.RequestId == "" {
-		return errMissingRequestID
+func (x *WakuMessageKeyValue) Validate() error {
+	if len(x.MessageHash) != 32 {
+		return errInvalidMessageHash
 	}
 
-	if x.RequestId != requestID {
-		return errRequestIDMismatch
+	if x.Message == nil {
+		return errMissingMessage
+	} else {
+		return x.Message.Validate()
 	}
-
-	if x.Response != nil {
-		return x.Response.Validate()
-	}
-
-	return nil
 }
